npm(eslint-config): add bugs entry
Summary:
Well, it's unlikely someone will need file a bug for this pkg but tools
are referring to GitHub issues which are disabled.

Not bumping the version because trivial.

Signed-off-by: Yongmin Hong <revi@omglol.email>

Test Plan: When new version is published, go to deps.dev and verify issues label is set to `revi.xyz/new`

Reviewers: O1 revi & automations, revi

Reviewed By: O1 revi & automations, revi

Differential Revision: https://issuetracker.revi.xyz/D416